% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete}
\alias{complete}
\title{A Notify Function}
\usage{
complete()
}
\description{
This function notifies you when the task is completed.
}
\details{
Description
}
\examples{
tic()
rnorm(10) + rnorm(20)
complete()
}
